+++ examples/cookie-banner-react/src/main.tsx
@@ -1,5 +1,8 @@
 import { createRoot } from "react-dom/client";
+import { installTCFStub } from "@probo/cookie-banner-tcf";
 import { App } from "./App";
+
+installTCFStub();
 
 createRoot(document.getElementById("root")!).render(
   <App />