+++ pkg/cmd/cookie-banner/view/view.go
@@ -45,6 +45,7 @@ query($id: ID!) {
       showBranding
       capabilities {
         resourceReporting
+        tcf
       }
       defaultLanguage
       publisherCountryCode
@@ -69,6 +70,7 @@ type viewResponse struct {
 		ShowBranding      bool    `json:"showBranding"`
 		Capabilities      struct {
 			ResourceReporting bool `json:"resourceReporting"`
+			TCF               bool `json:"tcf"`
 		} `json:"capabilities"`
 		DefaultLanguage      string `json:"defaultLanguage"`
 		PublisherCountryCode string `json:"publisherCountryCode"`
@@ -138,8 +140,10 @@ func NewCmdView(f *cmdutil.Factory) *cobra.Command {
 			_, _ = fmt.Fprintf(out, "%s%s\n", label.Render("State:"), v.State)
 			_, _ = fmt.Fprintf(out, "%s%d days\n", label.Render("Consent Expiry:"), v.ConsentExpiryDays)
 			_, _ = fmt.Fprintf(out, "%s%s\n", label.Render("Default Language:"), v.DefaultLanguage)
-			_, _ = fmt.Fprintf(out, "%s%s\n", label.Render("Publisher Country:"), v.PublisherCountryCode)
-			_, _ = fmt.Fprintf(out, "%s%d\n", label.Render("TCF CMP ID:"), v.TcfCmpId)
+			if v.Capabilities.TCF {
+				_, _ = fmt.Fprintf(out, "%s%s\n", label.Render("Publisher Country:"), v.PublisherCountryCode)
+				_, _ = fmt.Fprintf(out, "%s%d\n", label.Render("TCF CMP ID:"), v.TcfCmpId)
+			}
 			_, _ = fmt.Fprintf(out, "%s%t\n", label.Render("Resource Reporting:"), v.Capabilities.ResourceReporting)
 
 			_, _ = fmt.Fprintf(out, "%s%s\n", label.Render("Cookie Policy:"), v.CookiePolicyUrl)